@@ -11,14 +11,17 @@
     ],
     "license": "MIT",
     "require": {
-        "php": "^5.6 || ^7.0"
+        "php": "^7.3 || ^8.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "^5.5",
-        "chadicus/coding-standard": "^1.1",
-        "satooshi/php-coveralls": "^1.0"
+        "phpunit/phpunit": "^9.3",
+        "chadicus/coding-standard": "^2"
     },
     "autoload": {
         "psr-4": { "Chadicus\\": "src" }
+    },
+    "scripts": {
+        "lint": "vendor/bin/phpcs",
+        "test": "vendor/bin/phpunit"
     }
 }
